Improved  Performance
