+++ src/app/page.tsx
@@ -1,6 +1,3 @@
-"use client";
-import { useLayoutEffect, useState } from "react";
-import Loader from "@/components/Loader";
 import ContactForm from "@/components/Sections/ContactUs";
 import Section2 from "@/components/Sections/Section2";
 import ShopNowAt from "@/components/Sections/ShopNowAt";
@@ -14,18 +11,8 @@ import BestSellers from "@/components/Sections/bestSellers";
 import Landing_overlay from "@/components/Sections/landing_overlay";
 
 export default function Home() {
-  const [pageLoaded, setPageLoaded] = useState(false);
-
-  useLayoutEffect(() => {
-    setPageLoaded(document.readyState === "complete");
-  }, []);
-
-  if (!pageLoaded) {
-    return <Loader />;
-  }
-
   return (
-    <div className={`overflow-x-hidden ${!pageLoaded && "overflow-hidden"}`}>
+    <div className={`overflow-x-hidden `}>
       <Landing_overlay></Landing_overlay>
       <div className="px-1 pt-1 md:px-2 md:pt-2">
         <LandingPage></LandingPage>
