+++ src/types/types.ts
@@ -104,3 +104,22 @@ export const SignUpSchema = z
     message: "Passwords don't match",
     path: ["confirmPassword"],
   })
+
+
+  export interface BestSeller {
+    productid: string;
+    name: string;
+    price: number;
+    discount?: number;
+    category: string;
+    img : string
+}
+
+export interface NewArrival {
+  id : string;
+  name : string,
+  img : string,
+  price : number,
+  discount?: number,
+  category : string
+}
